apps/etl/requirements.txt
python-dotenv==1.0.1
awslambdaric==3.1.0
boto3==1.35.10
Pillow==10.0.0